refactor: Moved all configs to root
