--- config/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  env: {
-    node: true,
-  },
-  rules: {
-    '@typescript-eslint/no-var-requires': 'off',
-  },
-};
