+++ electron-builder.js
@@ -7,11 +7,11 @@ const buildVersion = `${now.getFullYear() - 2000}.${now.getMonth() + 1}.${now.ge
  */
 const config = {
   directories: {
-    output: 'dist/app',
+    output: 'dist',
     buildResources: 'buildResources',
   },
   files: [
-    'dist/source/**',
+    'packages/**/dist/**',
   ],
   extraMetadata: {
     version: buildVersion,
