+++ electron-vendors.config.js
@@ -1,7 +1,2 @@
 export const chrome = 89;
 export const node = 14;
-
-  // = {
-  // chrome: 89,
-  // node: 14,
-// };
