+++ external-packages.config.js
