+++ package.json
@@ -5,12 +5,12 @@
     "node": ">=v14.16.0",
     "npm": ">=7.6.3"
   },
-  "main": "dist/source/main/index.cjs.js",
+  "main": "packages/main/dist/index.cjs.cjs",
   "scripts": {
     "buildEnvTypes": "node ./bin/buildEnvTypes.js",
     "build": "node ./bin/build.js",
     "precompile": "cross-env MODE=production npm run build",
-    "compile": "electron-builder build --config ./config/electron-builder.js",
+    "compile": "electron-builder build --config electron-builder.js --dir",
     "pretest": "cross-env MODE=test npm run build",
     "test": "node ./tests/app.spec.js",
     "watch": "node ./bin/watch.js",
