+++ packages/renderer/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../config/tsconfig-base.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
